[FEATURE] Respect permissions in default actions
(cherry picked from commit 257f3331ae588d74b453f6f0e938752c3e5b2adc)

+++ Classes/Controller/NewsletterController.php
@@ -14,10 +14,9 @@ use In2code\Luxletter\Domain\Repository\UserRepository;
 use In2code\Luxletter\Domain\Service\PreviewUrlService;
 use In2code\Luxletter\Domain\Service\QueueService;
 use In2code\Luxletter\Domain\Service\ReceiverAnalysisService;
+use In2code\Luxletter\Events\AfterTestMailButtonClickedEvent;
 use In2code\Luxletter\Exception\ApiConnectionException;
 use In2code\Luxletter\Exception\AuthenticationFailedException;
-use In2code\Luxletter\Exception\InvalidUrlException;
-use In2code\Luxletter\Exception\MisconfigurationException;
 use In2code\Luxletter\Mail\TestMail;
 use In2code\Luxletter\Utility\BackendUserUtility;
 use In2code\Luxletter\Utility\ConfigurationUtility;
@@ -38,19 +37,9 @@ use TYPO3\CMS\Extbase\Persistence\Exception\InvalidQueryException;
 use TYPO3\CMS\Extbase\Persistence\Exception\UnknownObjectException;
 use TYPO3\CMS\Fluid\View\StandaloneView;
 
-/**
- * Class NewsletterController
- */
 class NewsletterController extends AbstractNewsletterController
 {
-    /**
-     * @return void
-     * @throws DBALException
-     * @throws ExceptionDbalDriver
-     * @throws ExceptionDbal
-     * @noinspection PhpUnused
-     */
-    public function dashboardAction(): void
+    public function dashboardAction(): ResponseInterface
     {
         $this->view->assignMultiple(
             [
@@ -69,24 +58,17 @@ class NewsletterController extends AbstractNewsletterController
                 'newsletters' => $this->newsletterRepository->findAll()->getQuery()->setLimit(10)->execute(),
             ]
         );
+
+        $this->addDocumentHeaderForNewsletterController();
+        return $this->defaultRendering();
     }
 
-    /**
-     * @return void
-     * @throws InvalidArgumentNameException
-     * @throws NoSuchArgumentException
-     */
     public function initializeListAction(): void
     {
         $this->setFilter();
     }
 
-    /**
-     * @param Filter $filter
-     * @return void
-     * @throws InvalidQueryException
-     */
-    public function listAction(Filter $filter): void
+    public function listAction(Filter $filter): ResponseInterface
     {
         $this->view->assignMultiple([
             'newsletters' => $this->newsletterRepository->findAllAuthorized(),
@@ -95,28 +77,23 @@ class NewsletterController extends AbstractNewsletterController
             'categories' => $this->categoryRepository->findAllLuxletterCategories(),
             'filter' => $filter,
         ]);
+
+        $this->addDocumentHeaderForNewsletterController();
+        return $this->defaultRendering();
     }
 
-    /**
-     * @param string $redirectAction
-     * @return void
-     * @throws StopActionException
-     */
-    public function resetFilterAction(string $redirectAction): void
+    public function resetFilterAction(string $redirectAction): ResponseInterface
     {
         BackendUserUtility::saveValueToSession('filter', $redirectAction, $this->getControllerName(), []);
-        $this->redirect($redirectAction);
+        return $this->redirect($redirectAction);
     }
 
-    /**
-     * @param Newsletter $newsletter
-     * @return void
-     * @throws ExceptionDbalDriver
-     * @throws InvalidConfigurationTypeException
-     * @throws DBALException
-     */
-    public function editAction(Newsletter $newsletter): void
+    public function editAction(Newsletter $newsletter): ResponseInterface
     {
+        if ($newsletter->canBeRead() === false) {
+            throw new AuthenticationFailedException('You are not allowed to see this record', 1709329205);
+        }
+
         $this->view->assignMultiple([
             'newsletter' => $newsletter,
             'configurations' => $this->configurationRepository->findAllAuthorized(),
@@ -127,33 +104,17 @@ class NewsletterController extends AbstractNewsletterController
         ]);
     }
 
-    /**
-     * @return void
-     * @throws NoSuchArgumentException
-     * @noinspection PhpUnused
-     */
     public function initializeUpdateAction(): void
     {
         $this->prepareArgumentsForPersistence();
     }
 
-    /**
-     * @param Newsletter $newsletter
-     * @return void
-     * @throws ApiConnectionException
-     * @throws ExceptionDbalDriver
-     * @throws ExtensionConfigurationExtensionNotConfiguredException
-     * @throws ExtensionConfigurationPathDoesNotExistException
-     * @throws IllegalObjectTypeException
-     * @throws InvalidConfigurationTypeException
-     * @throws InvalidUrlException
-     * @throws MisconfigurationException
-     * @throws SiteNotFoundException
-     * @throws StopActionException
-     * @throws UnknownObjectException
-     */
-    public function updateAction(Newsletter $newsletter): void
+    public function updateAction(Newsletter $newsletter): ResponseInterface
     {
+        if ($newsletter->canBeRead() === false) {
+            throw new AuthenticationFailedException('You are not allowed to see this record', 1709329247);
+        }
+
         $this->setBodytextInNewsletter($newsletter, $newsletter->getLanguage());
         if (ConfigurationUtility::isMultiLanguageModeActivated()) {
             $newsletter->setSubject(
@@ -169,14 +130,7 @@ class NewsletterController extends AbstractNewsletterController
         $this->redirect('list');
     }
 
-    /**
-     * @return void
-     * @throws InvalidConfigurationTypeException
-     * @throws ExceptionDbalDriver
-     * @throws DBALException
-     * @noinspection PhpUnused
-     */
-    public function newAction(): void
+    public function newAction(): ResponseInterface
     {
         $this->view->assignMultiple([
             'configurations' => $this->configurationRepository->findAllAuthorized(),
@@ -187,32 +141,17 @@ class NewsletterController extends AbstractNewsletterController
         ]);
     }
 
-    /**
-     * @return void
-     * @throws NoSuchArgumentException
-     * @noinspection PhpUnused
-     */
     public function initializeCreateAction(): void
     {
         $this->prepareArgumentsForPersistence();
     }
 
-    /**
-     * @param Newsletter $newsletter
-     * @return void
-     * @throws ApiConnectionException
-     * @throws ExceptionDbalDriver
-     * @throws ExtensionConfigurationExtensionNotConfiguredException
-     * @throws ExtensionConfigurationPathDoesNotExistException
-     * @throws IllegalObjectTypeException
-     * @throws InvalidConfigurationTypeException
-     * @throws InvalidUrlException
-     * @throws MisconfigurationException
-     * @throws SiteNotFoundException
-     * @throws StopActionException
-     */
-    public function createAction(Newsletter $newsletter): void
+    public function createAction(Newsletter $newsletter): ResponseInterface
     {
+        if ($newsletter->canBeRead() === false) {
+            throw new AuthenticationFailedException('You are not allowed to see this record', 1709329276);
+        }
+
         $languages = $this->pageRepository->getLanguagesFromOrigin($newsletter->getOrigin());
         foreach ($languages as $language) {
             $newsletterLanguage = clone $newsletter;
@@ -241,70 +180,45 @@ class NewsletterController extends AbstractNewsletterController
         $this->redirect('list');
     }
 
-    /**
-     * @param Newsletter $newsletter
-     * @return void
-     * @throws IllegalObjectTypeException
-     * @throws StopActionException
-     * @throws UnknownObjectException
-     */
-    public function disableAction(Newsletter $newsletter): void
+    public function disableAction(Newsletter $newsletter): ResponseInterface
     {
+        if ($newsletter->canBeRead() === false) {
+            throw new AuthenticationFailedException('You are not allowed to see this record', 1709329304);
+        }
+
         $newsletter->disable();
         $this->newsletterRepository->update($newsletter);
         $this->redirect('list');
     }
 
-    /**
-     * @param Newsletter $newsletter
-     * @return void
-     * @throws IllegalObjectTypeException
-     * @throws StopActionException
-     * @throws UnknownObjectException
-     * @noinspection PhpUnused
-     */
-    public function enableAction(Newsletter $newsletter): void
+    public function enableAction(Newsletter $newsletter): ResponseInterface
     {
+        if ($newsletter->canBeRead() === false) {
+            throw new AuthenticationFailedException('You are not allowed to see this record', 1709329338);
+        }
+
         $newsletter->enable();
         $this->newsletterRepository->update($newsletter);
         $this->redirect('list');
     }
 
-    /**
-     * @param Newsletter $newsletter
-     * @return void
-     * @throws IllegalObjectTypeException
-     * @throws StopActionException
-     * @throws DBALException
-     */
-    public function deleteAction(Newsletter $newsletter): void
+    public function deleteAction(Newsletter $newsletter): ResponseInterface
     {
+        if ($newsletter->canBeRead() === false) {
+            throw new AuthenticationFailedException('You are not allowed to see this record', 1709329345);
+        }
+
         $this->newsletterRepository->removeNewsletterAndQueues($newsletter);
         $this->addFlashMessage(LocalizationUtility::translate('module.newsletter.delete.message'));
         $this->redirect('list');
     }
 
-    /**
-     * Always pass a filter to receiverAction. If filter is given, save in session.
-     *
-     * @return void
-     * @throws NoSuchArgumentException
-     * @throws InvalidArgumentNameException
-     */
     public function initializeReceiverAction(): void
     {
         $this->setFilter();
     }
 
-    /**
-     * @param Filter $filter
-     * @return void
-     * @throws DBALException
-     * @throws ExceptionDbalDriver
-     * @throws InvalidQueryException
-     * @noinspection PhpUnused
-     */
-    public function receiverAction(Filter $filter): void
+    public function receiverAction(Filter $filter): ResponseInterface
     {
         $receiverAnalysisService = GeneralUtility::makeInstance(ReceiverAnalysisService::class);
         $users = $this->userRepository->getUsersByFilter($filter);
@@ -318,13 +232,6 @@ class NewsletterController extends AbstractNewsletterController
         );
     }
 
-    /**
-     * @param ServerRequestInterface $request
-     * @return ResponseInterface
-     * @throws DBALException
-     * @throws ExceptionDbalDriver
-     * @noinspection PhpUnused
-     */
     public function wizardUserPreviewAjax(ServerRequestInterface $request): ResponseInterface
     {
         $usergroups = GeneralUtility::intExplode(',', $request->getQueryParams()['usergroups'], true);
@@ -342,65 +249,53 @@ class NewsletterController extends AbstractNewsletterController
         return $response;
     }
 
-    /**
-     * @param ServerRequestInterface $request
-     * @return ResponseInterface
-     * @throws ApiConnectionException
-     * @throws AuthenticationFailedException
-     * @throws ExceptionDbalDriver
-     * @throws ExtensionConfigurationExtensionNotConfiguredException
-     * @throws ExtensionConfigurationPathDoesNotExistException
-     * @throws InvalidConfigurationTypeException
-     * @throws InvalidUrlException
-     * @throws MisconfigurationException
-     * @noinspection PhpUnused
-     */
     public function testMailAjax(ServerRequestInterface $request): ResponseInterface
     {
         if (BackendUserUtility::isBackendUserAuthenticated() === false) {
             throw new AuthenticationFailedException('You are not authenticated to send mails', 1560872725);
         }
-        $testMail = GeneralUtility::makeInstance(TestMail::class);
-        $status = $testMail->preflight(
-            $request->getQueryParams()['origin'],
-            $request->getQueryParams()['layout'],
-            (int)$request->getQueryParams()['configuration'],
-            $request->getQueryParams()['subject'],
-            $request->getQueryParams()['email']
-        );
+        $status = null;
+
+        /**
+         * This event can be used for sending the Test-email with external logic
+         * @see Documentation/Tech/Events.md
+         */
+        $event = GeneralUtility::makeInstance(AfterTestMailButtonClickedEvent::class, $request);
+        $this->eventDispatcher->dispatch($event);
+
+        if ($event->isTestMailIsSendExternal() === false) {
+            $testMail = GeneralUtility::makeInstance(TestMail::class);
+            $status = $testMail->preflight(
+                $request->getQueryParams()['origin'],
+                $request->getQueryParams()['layout'],
+                (int)$request->getQueryParams()['configuration'],
+                $request->getQueryParams()['subject'],
+                $request->getQueryParams()['email']
+            );
+        }
+        $responseData = [
+            'status' => $status ?? $event->getStatus(),
+        ];
+        if ($event->isTestMailIsSendExternal()) {
+            $responseData += $event->getStatusResponse();
+        }
         $response = ObjectUtility::getJsonResponse();
-        $response->getBody()->write(json_encode(['status' => $status]));
+        $response->getBody()->write(json_encode($responseData, JSON_THROW_ON_ERROR));
         return $response;
     }
 
-    /**
-     * @param ServerRequestInterface $request
-     * @return ResponseInterface
-     * @throws AuthenticationFailedException
-     * @throws ExceptionDbalDriver
-     * @throws ExtensionConfigurationExtensionNotConfiguredException
-     * @throws ExtensionConfigurationPathDoesNotExistException
-     * @throws MisconfigurationException
-     * @throws InvalidConfigurationTypeException
-     * @noinspection PhpUnused
-     */
     public function previewSourcesAjax(ServerRequestInterface $request): ResponseInterface
     {
         if (BackendUserUtility::isBackendUserAuthenticated() === false) {
             throw new AuthenticationFailedException('You are not authenticated to send mails', 1645707268);
         }
         $previewUrlService = GeneralUtility::makeInstance(PreviewUrlService::class);
-        $response = ObjectUtility::getJsonResponse();
         $content = $previewUrlService->get($request->getQueryParams()['origin'], $request->getQueryParams()['layout']);
+        $response = ObjectUtility::getJsonResponse();
         $response->getBody()->write(json_encode($content));
         return $response;
     }
 
-    /**
-     * @param ServerRequestInterface $request
-     * @return ResponseInterface
-     * @noinspection PhpUnused
-     */
     public function receiverDetailAjax(ServerRequestInterface $request): ResponseInterface
     {
         $userRepository = GeneralUtility::makeInstance(UserRepository::class);
@@ -419,5 +314,17 @@ class NewsletterController extends AbstractNewsletterController
             ['html' => $standaloneView->render()]
         ));
         return $response;
+    }
+
+    protected function addDocumentHeaderForNewsletterController(): void
+    {
+        $menuConfiguration = [
+            'dashboard' => LocalizationUtility::translate('layout.backend.link.actiondashboard'),
+            'list' => LocalizationUtility::translate('layout.backend.link.actionnewsletter'),
+        ];
+        if (ConfigurationUtility::isReceiverActionActivated()) {
+            $menuConfiguration['receiver'] = LocalizationUtility::translate('layout.backend.link.actionreceiver');
+        }
+        $this->addDocumentHeader($menuConfiguration);
     }
 }
